refactor(Movement): 도메인 위치 변경
