+++ src/main/java/domain/piece/Ma.java
@@ -2,6 +2,8 @@ package domain.piece;
 
 import domain.JanggiCoordinate;
 import domain.board.JanggiBoard;
+import domain.piece.movement.Movement;
+
 import java.util.List;
 
 public class Ma implements Piece {
