+++ src/main/java/domain/piece/movement/Movement.java
@@ -1,4 +1,4 @@
-package domain.piece;
+package domain.piece.movement;
 
 import domain.JanggiCoordinate;
 import domain.board.JanggiBoard;
